+++ db/migrate/20200405010736_create_day_of_shokihaiishokus.rb
@@ -3,11 +3,11 @@ class CreateDayOfShokihaiishokus < ActiveRecord::Migration[6.0]
     create_table :day_of_shokihaiishokus do |t|
       t.references :report, null: false, foreign_key: true
       t.integer :day, null: false
-      t.integer :e2
-      t.integer :fsh
-      t.integer :lh
-      t.integer :p4
-      t.integer :endometrial_thickness
+      t.float :e2
+      t.float :fsh
+      t.float :lh
+      t.float :p4
+      t.float :endometrial_thickness
 
       t.timestamps
     end
