+++ db/migrate/20200405010754_create_day_of_haibanhoishokus.rb
@@ -3,11 +3,11 @@ class CreateDayOfHaibanhoishokus < ActiveRecord::Migration[6.0]
     create_table :day_of_haibanhoishokus do |t|
       t.references :report, null: false, foreign_key: true
       t.integer :day, null: false
-      t.integer :e2
-      t.integer :fsh
-      t.integer :lh
-      t.integer :p4
-      t.integer :endometrial_thickness
+      t.float :e2
+      t.float :fsh
+      t.float :lh
+      t.float :p4
+      t.float :endometrial_thickness
 
       t.timestamps
     end
